Implement rendering sections partially
The rendered output was tested on a Hugo site using data templates. This
was done manually by writing a data template file with the output from
the renderer component.

Working with a private git repo in go modules is a PITA so I'm going to
refactor this multiple binary approach to a single binary with multiple
repos.