Add admin commands to add feeds to table
This change moves the feeds from the config file to the table. However
collection and rendering part of the flow still works off the config
file. This will be fixed in the next commit. The added commands are
addFeed and listFeeds.

This change also adds skeletal code to doing cleanup and generating
archive pages. This flow is not tested yet.

Database file name specification was also moved to the config file to
keep the CLI arguments simple.